+++ libsolidity/interface/CompilerStack.cpp
@@ -779,11 +779,11 @@ bool CompilerStack::compile(State _stopAfter)
 					}
 					catch (Error const& _error)
 					{
-						m_errorReporter.codeGenerationError(_error);
+						reportCodeGenerationError(_error, contract);
 					}
 					catch (UnimplementedFeatureError const& _error)
 					{
-						reportUnimplementedFeatureError(_error);
+						reportUnimplementedFeatureError(_error, contract);
 					}
 
 					if (m_errorReporter.hasErrors())
@@ -1614,7 +1614,7 @@ void CompilerStack::generateEVMFromIR(ContractDefinition const& _contract)
 	if (stack.hasErrors())
 	{
 		for (std::shared_ptr<Error const> const& error: stack.errors())
-			reportIRPostAnalysisError(error.get());
+			reportIRPostAnalysisError(error.get(), compiledContract.contract);
 		return;
 	}
 
@@ -2010,18 +2010,47 @@ experimental::Analysis const& CompilerStack::experimentalAnalysis() const
 	return *m_experimentalAnalysis;
 }
 
-void CompilerStack::reportUnimplementedFeatureError(UnimplementedFeatureError const& _error)
+void CompilerStack::reportUnimplementedFeatureError(
+	UnimplementedFeatureError const& _error,
+	ContractDefinition const* _contractDefinition
+)
 {
 	solAssert(_error.comment(), "Errors must include a message for the user.");
+	if (_error.sourceLocation().sourceName)
+		solAssert(m_sources.count(*_error.sourceLocation().sourceName) != 0);
 
-	m_errorReporter.unimplementedFeatureError(1834_error, _error.sourceLocation(), *_error.comment());
+	m_errorReporter.unimplementedFeatureError(
+		1834_error,
+		(_error.sourceLocation().sourceName || !_contractDefinition) ?
+			_error.sourceLocation() :
+			_contractDefinition->location(),
+		*_error.comment()
+	);
 }
 
-void CompilerStack::reportIRPostAnalysisError(Error const* _error)
+void CompilerStack::reportCodeGenerationError(Error const& _error, ContractDefinition const* _contractDefinition)
+{
+	solAssert(_error.type() == Error::Type::CodeGenerationError);
+	solAssert(_error.comment(), "Errors must include a message for the user.");
+	if (_error.sourceLocation() && _error.sourceLocation()->sourceName)
+		solAssert(m_sources.count(*_error.sourceLocation()->sourceName) != 0);
+	solAssert(_contractDefinition);
+
+	m_errorReporter.codeGenerationError(
+		_error.errorId(),
+		(_error.sourceLocation() && _error.sourceLocation()->sourceName) ?
+			*_error.sourceLocation() :
+			_contractDefinition->location(),
+		*_error.comment()
+	);
+}
+
+void CompilerStack::reportIRPostAnalysisError(Error const* _error, ContractDefinition const* _contractDefinition)
 {
 	solAssert(_error);
 	solAssert(_error->comment(), "Errors must include a message for the user.");
 	solAssert(!_error->secondarySourceLocation());
+	solAssert(_contractDefinition);
 
 	// Do not report Yul warnings and infos. These are only reported in pure Yul compilation.
 	if (!Error::isError(_error->severity()))
@@ -2032,7 +2061,7 @@ void CompilerStack::reportIRPostAnalysisError(Error const* _error)
 		_error->type(),
 		// Ignore the original location. It's likely missing, but even if not, it points at Yul source.
 		// CompilerStack can only point at locations in Solidity sources.
-		SourceLocation{},
+		_contractDefinition->location(),
 		*_error->comment()
 	);
 }