+++ test/libsolidity/smtCheckerTests/operators/shifts/shr_unused.sol
@@ -8,4 +8,4 @@ contract C {
 // SMTEngine: all
 // ----
 // Info 1391: CHC: 1 verification condition(s) proved safe! Enable the model checker option "show proved safe" to see all of them.
-// UnimplementedFeatureError 1834: Not yet implemented - FixedPointType.
+// UnimplementedFeatureError 1834: (0-80): Not yet implemented - FixedPointType.