+++ test/libsolidity/syntaxTests/nameAndTypeResolution/304_fixed_type_rational_int_conversion.sol
@@ -7,4 +7,4 @@ contract test {
 }
 // ----
 // Warning 2018: (20-104): Function state mutability can be restricted to pure
-// UnimplementedFeatureError 1834: Not yet implemented - FixedPointType.
+// UnimplementedFeatureError 1834: (0-106): Not yet implemented - FixedPointType.